# Conflicts:
#	AdminSubpages/create_plan.py
#	crisis_events.csv

+++ AdminSubpages/admin_end_event.py
@@ -1,7 +1,8 @@
 import tkinter as tk
 from tkinter import ttk
 import pandas as pd
-
+import datetime as dt
+from tkinter import messagebox
 
 class AdminEndEvent:
     def __init__(self, window, back_button_to_admin_main):
@@ -20,19 +21,19 @@ class AdminEndEvent:
         end_plan_frame.grid_rowconfigure(3, weight=1)
 
         # Labels
-        end_plan_title = tk.Label(end_plan_frame, text="Admin End Plan", font=('Helvetica', 16))
+        end_plan_title = tk.Label(end_plan_frame, text="End Plan", font=('Helvetica', 16))
         end_plan_title.grid(row=0, column=0, columnspan=2, rowspan=2, sticky="nsew", pady=5, padx=5)
 
-        end_plan_tree = ttk.Treeview(end_plan_frame)
-        end_plan_tree.grid(row=3, column=0, sticky="nsew", padx=10, pady=5)
+        self.end_plan_tree = ttk.Treeview(end_plan_frame)
+        self.end_plan_tree.grid(row=3, column=0, sticky="nsew", padx=10, pady=5)
 
         # CSV data
         csv_file = "crisis_events.csv"
         #csv_data = self.load_csv_data(csv_file)
-        self.upload_csv_data(end_plan_tree, csv_file)
+        self.upload_csv_data(self.end_plan_tree, csv_file)
 
         # Buttons
-        end_event_btn = tk.Button(self.window, text="End Event", command=lambda: self.delete_csv_data_entry(end_plan_tree, csv_file))
+        end_event_btn = tk.Button(self.window, text="End Event", command=lambda: self.deactivate_csv_data_entry(self.end_plan_tree, csv_file))
         end_event_btn.grid(row=6, column=0, padx=5, pady=10)
 
         # Back button
@@ -42,31 +43,58 @@ class AdminEndEvent:
 
     def upload_csv_data(self, tree, filename):
         data = pd.read_csv(filename)
+        selected_attributes = ['Camp ID', 'Crisis Type', 'Description', 'Country', 'Day', 'Month', 'Year', 'Status', 'End Date']
+        data = data[selected_attributes]
+        # Only want to see 'Active' plans:
+        data = data[data['Status'] != 'Inactive']
 
         tree.delete(*tree.get_children())
-        tree['columns'] = list(data.columns)
+        tree['columns'] = selected_attributes[0:7]
         tree.column("#0", width=0, stretch=tk.NO)
         tree.heading("#0", text="", anchor=tk.W)
 
-        for col in data.columns:
-            tree.column(col, anchor=tk.CENTER, width=80)
+        # set size of column for date
+        tree.column("Day", width=2, anchor=tk.CENTER)
+        tree.column("Month", width=3, anchor=tk.CENTER)
+        tree.column("Year", width=2, anchor=tk.CENTER)
+
+        for col in selected_attributes[0:7]:
+            if col != "Day" and col != "Month" and col != "Year": #added condition based on attribute
+                tree.column(col, anchor=tk.CENTER, width=80)
             tree.heading(col, text=col, anchor=tk.CENTER)
 
         for index, row in data.iterrows():
-            tree.insert("", tk.END, values=list(row), iid=str(index))
+            #row_vals = row[selected_attributes[0:7]]
+            tree.insert("", tk.END, values=list(row[selected_attributes[0:7]]), iid=str(index))
 
 
-    def delete_csv_data_entry(self, tree, filename):
-        selected_item = tree.selection()
-        if selected_item:
-            # Get index of the selected row
-            index = int(selected_item[0])
+    def deactivate_csv_data_entry(self, tree, filename):
+        selected_item = self.end_plan_tree.focus()
+        if not selected_item:
+            messagebox.showinfo("No selection", "Please select to end a plan")
+            return
+        if messagebox.askokcancel("End plan?", "Are you sure you want to deactivate this plan?\nYour plan will be shown as 'Inactive'"):
+            selected_item = tree.selection()
+            if selected_item:
+                # Get index of the selected row
+                index = int(selected_item[0])
 
-            # Load current CSV data, remove selected row, and save back to CSV
-            data = pd.read_csv(filename)
-            data = data.drop(index)
-            data.to_csv(filename, index=False)
+                data = pd.read_csv(filename)
 
-            # Refresh the Treeview display
-            self.upload_csv_data(tree, filename)
+                current_datetime = dt.datetime.now()
+                formatted_datetime = current_datetime.strftime("%Y-%m-%d %H:%M:%S")
+
+                # Load current CSV data, remove selected row, and save back to CSV
+                if index < len(data):
+                    data.at[index, 'Status'] = 'Inactive'  # update status to Inactive
+                    data.at[index, 'End Date'] = formatted_datetime  # update status to datetime now
+
+                    # Save the updated data back to the CSV file
+                    data.to_csv(filename, index=False)
+
+                    # Remove the selected row from the Treeview
+                    tree.delete(selected_item)
+
+                # Refresh the Treeview display
+                #self.upload_csv_data(tree, filename)
